http_auth: fix debug message on 64-bits architectures
The * modifier expects an int. A pointer difference is ptrdiff_t.

+++ src/misc/http_auth.c
@@ -294,7 +294,7 @@ void http_auth_ParseWwwAuthenticateHeader(
         const char *psz_end = strchr( psz_header, ' ' );
         if ( psz_end )
             msg_Warn( p_this, "Unknown authentication scheme: '%*s'",
-                      psz_end - psz_header, psz_header );
+                      (int)(psz_end - psz_header), psz_header );
         else
             msg_Warn( p_this, "Unknown authentication scheme: '%s'",
                       psz_header );
